+++ packages/studio-core/src/index.ts
@@ -85,5 +85,4 @@ export * from './lib/utilities/visualEffects';
 export * from './lib/navigation/searchIndex';
 export * from './lib/navigation/useApplicationTransitionStore';
 export * from './lib/themeTransitionEngine';
-
-
+export * from './lib/designTokens';